magento-engcom/magento#34 upgrade moved to the install script
