+++ app/code/Magento/Inventory/etc/module.xml
@@ -6,7 +6,7 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="Magento_Inventory" setup_version="2.0.4">
+    <module name="Magento_Inventory" setup_version="2.0.0">
         <sequence>
             <module name="Magento_InventoryApi"/>
         </sequence>
